+++ sea-orm-cli/src/commands.rs
@@ -1,4 +1,3 @@
-
 use clap::ArgMatches;
 use sea_orm_codegen::{EntityTransformer, OutputFile, WithSerde, NameResolver};
 use std::{error::Error, fmt::Display, fs, io::Write, path::Path, process::Command, str::FromStr};
@@ -22,6 +21,12 @@ pub async fn run_generate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dy
                     .with_test_writer()
                     .try_init();
             }
+
+            let max_connections = args
+                .value_of("MAX_CONNECTIONS")
+                .map(str::parse::<u32>)
+                .transpose()?
+                .unwrap();
 
             // The database should be a valid URL that can be parsed
             // protocol://username:password@host/database_name
@@ -99,9 +104,9 @@ pub async fn run_generate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dy
             let table_stmts = match url.scheme() {
                 "mysql" => {
                     use sea_schema::mysql::discovery::SchemaDiscovery;
-                    use sqlx::MySqlPool;
+                    use sqlx::MySql;
 
-                    let connection = MySqlPool::connect(url.as_str()).await?;
+                    let connection = connect::<MySql>(max_connections, url.as_str()).await?;
                     let schema_discovery = SchemaDiscovery::new(connection, database_name);
                     let schema = schema_discovery.discover().await;
                     schema
@@ -114,9 +119,9 @@ pub async fn run_generate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dy
                 }
                 "sqlite" => {
                     use sea_schema::sqlite::SchemaDiscovery;
-                    use sqlx::SqlitePool;
+                    use sqlx::Sqlite;
 
-                    let connection = SqlitePool::connect(url.as_str()).await?;
+                    let connection = connect::<Sqlite>(max_connections, url.as_str()).await?;
                     let schema_discovery = SchemaDiscovery::new(connection);
                     let schema = schema_discovery.discover().await?;
                     schema
@@ -129,10 +134,10 @@ pub async fn run_generate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dy
                 }
                 "postgres" | "postgresql" => {
                     use sea_schema::postgres::discovery::SchemaDiscovery;
-                    use sqlx::PgPool;
+                    use sqlx::Postgres;
 
                     let schema = args.value_of("DATABASE_SCHEMA").unwrap_or("public");
-                    let connection = PgPool::connect(url.as_str()).await?;
+                    let connection = connect::<Postgres>(max_connections, url.as_str()).await?;
                     let schema_discovery = SchemaDiscovery::new(connection, schema);
                     let schema = schema_discovery.discover().await;
                     schema
@@ -172,6 +177,17 @@ pub async fn run_generate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dy
     };
 
     Ok(())
+}
+
+async fn connect<DB>(max_connections: u32, url: &str) -> Result<sqlx::Pool<DB>, Box<dyn Error>>
+where
+    DB: sqlx::Database,
+{
+    sqlx::pool::PoolOptions::<DB>::new()
+        .max_connections(max_connections)
+        .connect(url)
+        .await
+        .map_err(Into::into)
 }
 
 pub fn run_migrate_command(matches: &ArgMatches<'_>) -> Result<(), Box<dyn Error>> {
@@ -260,8 +276,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use clap::AppSettings;
     use crate::cli;
+    use clap::AppSettings;
 
     #[test]
     #[should_panic(