+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Pending
+
+### Enhancements
+* Add max_connections option to CLI https://github.com/SeaQL/sea-orm/pull/670
+
 ## 0.7.1 - 2022-03-26
 
 * Fix sea-orm-cli error